requirements.txt
python-dotenv==1.0.0
discord.py==2.3.1
asyncpg==0.27.0
fastapi==0.95.1
uvicorn[standard]
motor==3.1.2
httpx==0.24.0
aiofiles==23.1.0
prettytable==3.7.0
pytz==2023.3